requirements.txt
#Python 3.9, Tensorflow 2.8.0, Cuda 11.2 cudnn: 8.1.1
#1)Update pip installer with latest packages. Run terminal as admin to run next line.
# pip install --upgrade pip
#2)install requirements file
# pip install -r requirements.txt
#3)Install tensorflow examples
#pip install -q git+https://github.com/tensorflow/examples.git
#4)Check if using GPUs after install
# import tensorflow as tf
#tf.test.is_gpu_available(cuda_only=False, min_cuda_compute_capability=None)

#NOTE: tf.keras.layers.Randomxxx and by extension keras-cv do not work due to a bug in TF9, 10 and 11.
#Bugs would eventually be fixed but in the meantime use following work around:

tensorflow-gpu==2.8.0.
opencv-contrib-python== 4.6.0.66
keras-cv
imageio==2.23.0
mkl-fft
pandas==1.5.2
scikit-learn==1.2.0
XlsxWriter==3.0.3
jsonschema==4.17.3
sklearn==0.0
openpyxl==3.0.10
matplotlib==3.6.0
keras-applications==1.0.8
psutil==5.9.4
shapely
keras-tuner
keyboard
tensorflow_datasets


************************************************
tensorflow-gpu==2.5.2
opencv-contrib-python==3.4.2.17
imageio==2.13.3
mkl-fft==1.3.1
pandas==1.3.4
scikit-learn==1.0.1
XlsxWriter==3.0.2
jsonschema==4.2.1
sklearn==0.0
openpyxl==3.0.9
matplotlib==3.5.0
keras-applications==1.0.8
psutil==5.9.0
shapely
Augmentor
Augmentor --upgrade
keras-tuner
keyboard